requirements.txt
asgiref==3.3.4
cached-property==1.5.2
click==8.0.1
colorama==0.4.4
dataclasses==0.8
fastapi==0.65.2
flask
h11==0.12.0
h5py==3.1.0
importlib-metadata==4.5.0
jinja2
Keras==2.4.3
numpy==1.19.5
opencv-python==4.5.2.54
Pillow==8.2.0
pydantic==1.8.2
python-multipart==0.0.5
PyYAML==5.4.1
scipy==1.5.4
six==1.16.0
starlette==0.14.2
tensorflow
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.4.1
